Move bundled sqlite3 to vendors/.
