+++ tsk/hashdb/sqlite_hdb.cpp
@@ -15,7 +15,7 @@
 #ifdef HAVE_LIBSQLITE3
   #include <sqlite3.h>
 #else
-  #include "../auto/sqlite3.h"
+  #include "sqlite3.h"
 #endif
 
 /**
