+++ vendors/sqlite3.h
